Update db schema and ticket_dispenser_gem to solve rspec related errors

+++ db/migrate/20250730205439_remove_index_articles_courses_on_course_id.rb
@@ -1,5 +1,6 @@
 class RemoveIndexArticlesCoursesOnCourseId < ActiveRecord::Migration[7.0]
   def change
-    remove_index :articles_courses, name: :index_articles_courses_on_course_id
+    if index_exists?(:articles_courses, name: :index_articles_courses_on_course_id)
+      remove_index :articles_courses, name: :index_articles_courses_on_course_id
   end
 end
